libfshelp: type check messages using the full mach_msg_type_t so that it works on x86_64.
Message-ID: <ZVlHm2xuVaS8SItY@jupiter.tail36e24.ts.net>

+++ libfshelp/start-translator-long.c
@@ -110,7 +110,7 @@ service_fsys_startup (fshelp_open_fn_t underlying_open_fn, void *cookie,
     {
       union
       {
-        uint32_t word;
+        uintptr_t word;
 	mach_msg_type_t type;
       } t, c;
       t.type = *type;
